fix: 6372 GE image download (#6375)
