+++ frontend/next.config.js
@@ -22,6 +22,8 @@ const CROSS_REF_URL = "https://api.crossref.org";
 
 const DATADOG_URL = "browser-intake-datadoghq.com";
 
+const GOOGLE_FONTS_URL = "https://fonts.gstatic.com";
+
 const SCRIPT_SRC = [
   "'self'",
   "'wasm-unsafe-eval'",
@@ -46,9 +48,10 @@ const defaultSecureHeaders = {
         configs.CENSUS_MODELS_DATA_URL,
         CROSS_REF_URL,
         DATADOG_URL,
+        GOOGLE_FONTS_URL,
       ],
       defaultSrc: ["'self'", HUBSPOT_FORMS_URL],
-      fontSrc: ["'self'", "https://fonts.gstatic.com"],
+      fontSrc: ["'self'", GOOGLE_FONTS_URL],
       formAction: ["'self'", HUBSPOT_FORMS_URL],
       frameAncestors: ["'none'"],
       // 4513(thuang): Comment out frameSrc for now until we figure out a compliant way to embed
