+++ frontend/src/views/WheresMyGeneV2/components/GeneSearchBar/components/SaveExport/index.tsx
@@ -70,6 +70,7 @@ import {
 } from "src/views/WheresMyGeneV2/components/HeatMap/utils";
 import { CHART_PADDING_PX } from "src/views/WheresMyGeneV2/components/HeatMap/style";
 import { StyledButtonIcon } from "../ShareButton/style";
+import { getCurrentDate } from "src/views/WheresMyGeneV2/components/GeneSearchBar/components/SaveExport/utils";
 
 let heatmapContainerScrollTop: number | undefined;
 
@@ -749,14 +750,4 @@ function download_({
     setEchartsRendererMode("canvas");
     setDownloadStatus({ isLoading: false });
   };
-}
-
-// Gets the date in mmddyy format
-export function getCurrentDate() {
-  const today = new Date();
-  const month = (today.getMonth() + 1).toString().padStart(2, "0");
-  const day = today.getDate().toString().padStart(2, "0");
-  const year = today.getFullYear().toString().slice(-2);
-
-  return month + day + year;
 }
